Levenshtein's function
I create the distance matrix with the minimum cost (between detetion, insertion, or substitution) in each component ij and the matrix of the letters with the initial of the minimization process (deletion, insertion or substitution) associated to each component in the distance matrix. Then, I do the backtracking.
Finally, the DNA function reads a file and find the best line of the file where the pattern is better represented.